LogAggregationFilesBuilder: Move to separate class
